test(google-docs): move entry title helper coverage into dedicated spec
