+++ apps/google-docs/test/utils/overviewEntryList.test.ts
@@ -1,10 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import type { CompletedWorkflowPayload } from '../../src/types';
 import type { ContentTypeDisplayInfo } from '../../src/services/contentTypeService';
-import {
-  buildEntryList,
-  buildEntryListFromEntryBlockGraph,
-} from '../../src/utils/overviewEntryList';
+import { buildEntryList } from '../../src/utils/overviewEntryList';
 
 describe('buildEntryList', () => {
   it('omits entries without tempId', () => {
@@ -118,48 +115,5 @@ describe('buildEntryList', () => {
     const rows = buildEntryList(payload, contentTypeDisplayInfoById, 'en-US');
     expect(rows[0].contentTypeName).toBe('Page (space name)');
     expect(rows[0].entryTitle).toBe('Event Detail');
-  });
-});
-
-describe('buildEntryListFromEntryBlockGraph', () => {
-  const contentTypes = [
-    {
-      sys: { id: 'article' },
-      name: 'Article',
-      displayField: 'title',
-      fields: [],
-    },
-  ];
-
-  const createTextSourceRef = (text: string) => ({
-    type: 'blockText' as const,
-    blockId: 'block-1',
-    start: 0,
-    end: text.length,
-    flattenedRuns: [{ text, start: 0, end: text.length }],
-  });
-
-  it('extracts the entry title from the display field mapping', () => {
-    const rows = buildEntryListFromEntryBlockGraph(
-      [
-        {
-          tempId: 'entry-1',
-          contentTypeId: 'article',
-          fieldMappings: [
-            {
-              fieldId: 'title',
-              fieldType: 'Symbol',
-              sourceRefs: [createTextSourceRef('Entry title')],
-              confidence: 0.9,
-            },
-          ],
-        },
-      ],
-      contentTypes
-    );
-
-    expect(rows).toHaveLength(1);
-    expect(rows[0].contentTypeName).toBe('Article');
-    expect(rows[0].entryTitle).toBe('Entry title');
   });
 });
